requirements.txt
aiogram==3.11.0
aiosqlite==0.20.0
python-dotenv==1.0.0
Pillow==10.4.0